move resolutions_to_tasks() to core.utils
This is moving the method resolutions_to_tasks to core.utils just for a
better organization.
