+++ avocado/core/job.py
@@ -26,65 +26,19 @@ import sys
 import tempfile
 import time
 import traceback
-import uuid
 
 from ..utils import astring, path, process, stacktrace
 from ..utils.data_structures import CallbackRegister, time_to_seconds
 from . import (data_dir, dispatcher, exceptions, exit_codes, jobdata,
-               loader, nrunner, output, result, tags, varianter, version)
+               loader, output, result, tags, varianter, version)
 from .job_id import create_unique_job_id
 from .future.settings import settings
 from .output import LOG_JOB, LOG_UI, STD_OUTPUT
-from .resolver import ReferenceResolutionResult, resolve
-from .tags import filter_test_tags_runnable
 from .test import DryRunTest
+from .utils import resolutions_to_tasks
+from .resolver import resolve
 
 _NEW_ISSUE_LINK = 'https://github.com/avocado-framework/avocado/issues/new'
-
-
-def resolutions_to_tasks(resolutions, config):
-    """
-    Transforms resolver resolutions into tasks
-
-    A resolver resolution
-    (:class:`avocado.core.resolver.ReferenceResolution`) contains
-    information about the resolution process (if it was successful
-    or not) and in case of successful resolutions a list of
-    resolutions.  It's expected that the resolution are
-    :class:`avocado.core.nrunner.Runnable`.
-
-    This method transforms those runnables into Tasks
-    (:class:`avocado.core.nrunner.Task`), which will include a status
-    reporting URI.  It also performs tag based filtering on the
-    runnables for possibly excluding some of the Runnables.
-
-    :param resolutions: possible multiple resolutions for multiple
-                        references
-    :type resolutions: list of :class:`avocado.core.resolver.ReferenceResolution`
-    :param config: job configuration
-    :type config: dict
-    :returns: the resolutions converted to tasks
-    :rtype: list of :class:`avocado.core.nrunner.Task`
-    """
-
-    tasks = []
-    filter_by_tags = config.get("filter.by_tags.tags")
-    include_empty = config.get("filter.by_tags.include_empty")
-    include_empty_key = config.get('filter.by_tags.include_empty_key')
-    status_server = config.get('nrun.status_server.listen')
-    for resolution in resolutions:
-        if resolution.result != ReferenceResolutionResult.SUCCESS:
-            continue
-        for runnable in resolution.resolutions:
-            if filter_by_tags:
-                if not filter_test_tags_runnable(runnable,
-                                                 filter_by_tags,
-                                                 include_empty,
-                                                 include_empty_key):
-                    continue
-            tasks.append(nrunner.Task(str(uuid.uuid1()), runnable,
-                                      [status_server]))
-    return tasks
 
 
 def register_job_options():
